pyproject.toml
[tool.poetry]
name = "ce-portal"
version = "0.1.0"
description = ""
authors = ["a.zinoviev"]
packages = [{include = "ce_portal"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
arrow = "^1.2.3"
django-mptt = "^0.14.0"
django-taggit = "^3.0.0"
django-modelcluster = "^6.0"
six = "^1.16.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
pillow = "^9.2.0"
django-ckeditor = "^6.5.1"
django-crispy-forms = "^1.14.0"
django-admin-interface = "^0.22.1"
django-solo = "^2.0.0"
sorl-thumbnail = "^12.9.0"
django-taggit-labels = "^0.7.0"
django-taggit-anywhere = "^1.0.0"
django-taggit-helpers = "^0.1.4"
django-birthday = "^0.1.4"
django-vote = "^2.3.0"
django-editorjs = "^0.2.1"
django-autoslug = "^1.9.8"
django-editorjs-fields = "^0.2.7"
django-view-breadcrumbs = "^2.4.1"
pytils = "^0.4.1"
django-threadedcomments = "^2.0"
drf-yasg = "^1.21.4"
django-widget-tweaks = "^1.4.12"
django-model-utils = "^4.2.0"
django-htmx = "^1.12.2"
django-colorfield = "^0.7.2"
django-taggit-templatetags = "^0.2.5"
django-filer = "^2.2.3"
django-compressor = "^4.1"
django-libsass = "^0.9"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
uvicorn = "^0.19.0"
psycopg2 = "^2.9.5"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
